# docs/url-shortener/URLShortener_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\n***IMPORTANT! READ THIS BEFORE TEST***\n\nWhen entering the token in Swagger Authorization:\n**\"Bearer {YOUR-ACCESS-TOKEN}\"**\nMake sure to include Bearer before your access token\n\nAlways open **\"DevTools (F12) -\u003e Application -\u003e Cookies\"**\nFor getting the latest Refresh Token and Session ID cookie in real time\nWorks with HttpOnly cookies, which are protected from JavaScript access.",
        "title": "URL Shortener Project API Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/url-shortener",
    "paths": {
        "/admin/shortens": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All User Shortens Code Data from database with Paginated List",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get All User Shortens Code Paginated List (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (Default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (Default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Shortens Code Successfully",
                        "schema": {
                            "$ref": "#/definitions/shortener.ShortensListSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "403": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "404": {
                        "description": "Code Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
        "/admin/shortens/{shid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Shortener Code Data from database by Shorten ID(shid)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get User Shortener Code by Shorten ID(shid) (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "sub",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Shorten ID",
                        "name": "shid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Shortener Code Successfully",
                        "schema": {
                            "$ref": "#/definitions/shortener.ShortenerSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "403": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "404": {
                        "description": "Code Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
        "/admin/users/{sub}/shortens": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get User Shortens Code Paginated List from database by Shorten ID(shid)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get User Shortens Code Paginated List by Shorten ID(shid) (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "sub",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page (Default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (Default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Shortens Code Successfully",
                        "schema": {
                            "$ref": "#/definitions/shortener.ShortenerSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "403": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "404": {
                        "description": "Code Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
        "/s": {
            "post": {
                "description": "Create Shorten Code from OriginalURL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener Public"
                ],
                "summary": "Create Shorten Code Public",
                "parameters": [
                    {
                        "description": "Create Shorten Public Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shortener.CreateShortenPublicRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Short URL Successfully",
                        "schema": {
                            "$ref": "#/definitions/shortener.ShortenerSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid Body Requesy",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
        "/s/{code}": {
            "get": {
                "description": "Redirect to Original URL from Shorten Code in Redis Cache",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener Public"
                ],
                "summary": "Redirect to Original URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shorten Code from Generate Shorten Link",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecting to Original URL",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "404": {
                        "description": "Shorten Code Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
        "/shortens/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Shortens Code Data from database with Paginated List",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener Authorized"
                ],
                "summary": "Get All Shortens Code Paginated List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (Default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (Default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Shortens Code Successfully",
                        "schema": {
                            "$ref": "#/definitions/shortener.ShortensListSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "404": {
                        "description": "Code Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Shorten Code from OriginalURL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener Authorized"
                ],
                "summary": "Create Shorten Code Authorized",
                "parameters": [
                    {
                        "description": "Create Shorten Authorized Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shortener.CreateShortenAuthorizedRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Short URL Successfully",
                        "schema": {
                            "$ref": "#/definitions/shortener.ShortenerSwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Body Requesy",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pkg.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "shortener.CreateShortenAuthorizedRequest": {
            "type": "object",
            "required": [
                "original_url"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                }
            }
        },
        "shortener.CreateShortenPublicRequest": {
            "type": "object",
            "required": [
                "original_url"
            ],
            "properties": {
                "original_url": {
                    "type": "string"
                }
            }
        },
        "shortener.PaginatedShortenResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/shortener.PaginationMeta"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shortener.ShortenResponse"
                    }
                }
            }
        },
        "shortener.PaginationMeta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "shortener.ShortenerResponseData": {
            "type": "object",
            "properties": {
                "shorten": {
                    "$ref": "#/definitions/shortener.ShortenResponse"
                }
            }
        },
        "shortener.ShortenerSwaggerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/shortener.ShortenerResponseData"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "shortener.ShortensListSwaggerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/shortener.ShortensResponseData"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "shortener.ShortensResponseData": {
            "type": "object",
            "properties": {
                "shortens": {
                    "$ref": "#/definitions/shortener.PaginatedShortenResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: **\"Bearer {YOUR_ACCESS_TOKEN}\"**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}